requirements.txt
matplotlib==3.7.5
moviepy==1.0.3
neptune==1.13.0
numpy==1.22.4
opencv-python==4.5.5.64
pandas==1.3.5
pyrender==0.1.45
PyYAML==6.0
ray[tune]==2.10.0
seaborn==0.11.2
scikit-learn==1.1.1
smplx==0.1.28
tqdm==4.63.0
